gas_demand_response/rl_training/requirements.txt
# local package
-e .

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1

pandas==1.4.2
numpy==1.22.3
pyyaml==6.0
tqdm==4.64.0
wandb==0.12.18
plotly==5.8.2
scikit-learn==1.0.2
setuptools==61.2.0

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu